Move TestNG XML and properties files to their respective scopes
