+++ .templates/eclipseLaunches/All tests.launch.xml
@@ -17,8 +17,8 @@
 <booleanAttribute key="org.testng.eclipse.PREFIX_VM_ARGS_FROM_POM" value="true"/>
 <stringAttribute key="org.testng.eclipse.PROTOCOL" value="object"/>
 <listAttribute key="org.testng.eclipse.SUITE_TEST_LIST">
-<listEntry value="src/test/testng-component.xml"/>
-<listEntry value="src/test/testng-e2e.xml"/>
+<listEntry value="src/test/resources/testng-component.xml"/>
+<listEntry value="src/e2e/resources/testng-e2e.xml"/>
 </listAttribute>
 <intAttribute key="org.testng.eclipse.TYPE" value="3"/>
 <booleanAttribute key="org.testng.eclipse.VERBOSE" value="false"/>
