+++ .templates/ideaRunConfigurations/E2E_Tests.xml
@@ -4,7 +4,7 @@
     <module name="" />
     <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="true" />
     <option name="ALTERNATIVE_JRE_PATH" value="1.8" />
-    <option name="SUITE_NAME" value="$PROJECT_DIR$/src/test/testng-e2e.xml" />
+    <option name="SUITE_NAME" value="$PROJECT_DIR$/src/e2e/resources/testng-e2e.xml" />
     <option name="PACKAGE_NAME" value="" />
     <option name="MAIN_CLASS_NAME" value="" />
     <option name="METHOD_NAME" value="" />
