+++ build.gradle
@@ -115,7 +115,7 @@ task createConfigs {
         def templatesToCopy = [
             "gradle.template.properties",
             "src/main/resources/build.template.properties",
-            "src/test/resources/test.template.properties",
+            "src/e2e/resources/test.template.properties",
             "src/main/webapp/WEB-INF/appengine-web.template.xml"
         ]
         templatesToCopy.findAll {
@@ -511,7 +511,7 @@ task e2eTests {
     def isLastRetry = id == numOfTestRetries + 1
     task "componentTestTry${id}"(type: Test) {
         useTestNG()
-        options.suites isFirstTry ? "src/test/testng-component.xml" : "build/reports/component-test-try-${id - 1}/testng-failed.xml"
+        options.suites isFirstTry ? "src/test/resources/testng-component.xml" : "build/reports/component-test-try-${id - 1}/testng-failed.xml"
         options.outputDirectory = file("build/reports/component-test-try-${id}")
         options.useDefaultListeners = true
         ignoreFailures = !isLastRetry
@@ -539,7 +539,7 @@ task e2eTests {
 
     task "e2eTestTry${id}"(type: Test) {
         useTestNG()
-        options.suites isFirstTry ? "src/test/testng-e2e.xml" : "build/reports/e2e-test-try-${id - 1}/testng-failed.xml"
+        options.suites isFirstTry ? "src/e2e/resources/testng-e2e.xml" : "build/reports/e2e-test-try-${id - 1}/testng-failed.xml"
         options.outputDirectory = file("build/reports/e2e-test-try-${id}")
         options.useDefaultListeners = true
         ignoreFailures = !isLastRetry
