+++ src/e2e/resources/testng-e2e.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
 <suite name="e2e-tests" parallel="tests" thread-count="2">
+    <listeners>
+        <listener class-name="teammates.e2e.util.PriorityInterceptor" />
+    </listeners>
     <test name="e2e-tests">
         <packages>
             <package name="teammates.e2e.cases.util" />
